+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CTGen 1.1
 ================
 
 - port to 1.21.4
+- add handlers to SyncMapPacket
 
 Crafted Terrain Generation
 ================